pyproject.toml
[tool.poetry]
name = "fine_tune_llm"
version = "0.1.0"
description = ""
authors = ["ccirelli2 <ccirelli2@gmail.com>"]
readme = "README.md"
packages = [{include='src'}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ollama = "^0.3.0"
transformers = "^4.43.3"
torch = "^2.4.0"
langchain-text-splitters = "^0.2.2"
langchain = "^0.2.11"
fastembed = "^0.3.4"
pytesseract = "^0.3.10"
pillow = "^10.4.0"
pdf2image = "^1.17.0"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
langchain-community = "^0.2.10"
huggingface-hub = "^0.24.2"
mysql-connector-python = "^9.0.0"
chromadb = "^0.5.5"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
streamlit = "^1.37.0"
prefect = "^2.19.9"
python-decouple = "^3.8"
GitPython = "^3.1.43"
numpy = ">=1.22.5, <2.0.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
jupyter = "^1.0.0"
lab = "^8.2"
protobuf = "3.20.3"
chardet = "^5.2.0"
aiofiles = "^24.1.0"
python-dotenv = "^1.0.1"
watchdog = "^4.0.1"
sqlalchemy = "^2.0.31"
accelerate = "^0.33.0"
llama-index-llms-ollama = "^0.2.2"
llama-index = "^0.10.61"
tomlkit = "^0.13.0"
black = "^24.8.0"
aiomysql = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
